app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prepper-cli"
version = "0.1.0"
description = "CLI and shared OpenRouter client for Prepper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["openai", "python-dotenv"]

[project.scripts]
prepper-cli = "prepper_cli.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
prepper_cli = ["prompts/*.md"]
